G9-api/UserService/generatedCode/requirements.txt
Flask==2.2.3
flask_expects_json==1.7.0
jsonschema==4.17.3
python-dotenv==1.0.0
redis==4.5.1
requests==2.28.2
waitress==2.1.2
